Converted outputs to match publisher list, removed depth post processing
